+++ Communication/Packets/Incoming/Moderation/OpenHelpToolEvent.cs
@@ -1,13 +1,18 @@
 ﻿using Plus.Communication.Packets.Outgoing.Moderation;
 using Plus.HabboHotel.GameClients;
+using Plus.HabboHotel.Guides;
 
 namespace Plus.Communication.Packets.Incoming.Moderation;
 
 internal class OpenHelpToolEvent : IPacketEvent
 {
-    public Task Parse(GameClient session, IIncomingPacket packet)
+    private readonly IGuideService _guideService;
+
+    public OpenHelpToolEvent(IGuideService guideService) => _guideService = guideService;
+
+    public async Task Parse(GameClient session, IIncomingPacket packet)
     {
         session.Send(new OpenHelpToolComposer());
-        return Task.CompletedTask;
+        await _guideService.SendToolState(session);
     }
 }